implemented generic types

+++ src/CoreComputing.jl
@@ -310,7 +310,7 @@ function project_particles!(
         xproj = dot(pⱼ.coordinates - cellᵢ.center, Δc)
         if abs(xproj) <= margin
             iproj += 1
-            projected_particles[iproj] = ProjectedParticle(pⱼ.index, Float32(xproj), pⱼ.coordinates) 
+            projected_particles[iproj] = ProjectedParticle(pⱼ.index, xproj, pⱼ.coordinates) 
         end
     end
     pp = @view(projected_particles[1:iproj])
